+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - linear source inversion *lsrcinv* in *nessi.signal*
+- dispersion module *nessi.modeling.dispersion*
+- Rayleigh wave velocity estimation *vrayleigh* in *nessi.modeling.dispersion*
 
 ## [0.2.2] - 2018-12-05